requirements.txt
flask
python-dotenv
pypdf
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
faiss-cpu
datasets
ragas
